FIXES #3 Windows compatibility
Use identify -list font for finding the absolute paths to the fonts. Parses the output from ImageMagick and searches by name or ttf filename.

DejaVuSans is missing on windows, we also check if a font exists, and do not add a specific font if it is missing. All warnings fixed.