First Ideas on New API
Idea: The script is fine, but to use the API from your own scripts or rails, you hit a few bumps. It would be much more natural to create a video Object first. On the video Object we can grab single frames, create contact sheets and query about the video file.

* New Video Object to get metadata about the video
* Rename thumbnail to more generic frame
* Adapted FFMPEG parser to use -show_format -show_streams (easier parsing, more information)
** Can in principle handle more than one video or audio streams